Revert "Merge branch 'main' into mult_img_dropdown_v2"
This reverts commit 1da22ffd439836455fc3e5a94d68633197330dcd, reversing
changes made to b89e6433286d829839aa8850737e00755758ef67.
