+++ apps/dashboard/src/app/AppHeader.tsx
@@ -6,12 +6,12 @@ import {
   ICommandBarItemProps,
   IContextualMenuItem
 } from "@fluentui/react";
+import { Language } from "@responsible-ai/localization";
 import {
   featureFlights,
   featureFlightSeparator,
   parseFeatureFlights
-} from "@responsible-ai/core-ui";
-import { Language } from "@responsible-ai/localization";
+} from "@responsible-ai/model-assessment";
 import React from "react";
 
 import { applications, IApplications } from "./applications";
