+++ libs/core-ui/src/index.ts
@@ -55,14 +55,12 @@ export * from "./lib/util/calculateBoxData";
 export * from "./lib/util/calculateConfusionMatrixData";
 export * from "./lib/util/calculateLineData";
 export * from "./lib/util/MultilabelStatisticsUtils";
-export * from "./lib/util/ObjectDetectionStatisticsUtils";
 export * from "./lib/util/StatisticsUtils";
 export * from "./lib/util/string";
 export * from "./lib/util/toScientific";
 export * from "./lib/util/ITelemetryEvent";
 export * from "./lib/util/TelemetryEventName";
 export * from "./lib/util/generateDefaultChartAxes";
-export * from "./lib/util/datasetUtils/getColumnRanges";
 export * from "./lib/components/Announce";
 export * from "./lib/components/OverallMetricChart";
 export * from "./lib/components/AxisConfig";
@@ -77,7 +75,6 @@ export * from "./lib/components/MissingParametersPlaceholder";
 export * from "./lib/components/LabelWithCallout";
 export * from "./lib/components/NoData";
 export * from "./lib/components/SVGToolTip";
-export * from "./lib/FeatureFlights";
 export * from "./lib/Interfaces/ComparisonTypes";
 export * from "./lib/Interfaces/ExplanationInterfaces";
 export * from "./lib/Interfaces/IExplanationContext";
