+++ libs/core-ui/src/lib/Cohort/CohortBasedComponent.ts
@@ -8,9 +8,8 @@ import { IDataset } from "../Interfaces/IDataset";
 import { IFilter, ICompositeFilter } from "../Interfaces/IFilter";
 import { JointDataset } from "../util/JointDataset";
 
-import { Cohort } from "./Cohort";
+import { Cohort, CohortSource } from "./Cohort";
 import { MetricCohortStats } from "./CohortStats";
-import { CohortSource } from "./Constants";
 import { ErrorCohort } from "./ErrorCohort";
 
 export interface ICohortBasedComponentProps {
