+++ libs/core-ui/src/lib/Cohort/Constants.ts
@@ -23,11 +23,3 @@ export class Metrics {
   public static RecallScore = "Recall score";
   public static SampleSize = "Sample size";
 }
-
-export enum CohortSource {
-  None = "None",
-  TreeMap = "Tree map",
-  HeatMap = "Heat map",
-  ManuallyCreated = "Manually created",
-  Prebuilt = "Prebuilt"
-}
