--- libs/core-ui/src/lib/DatasetCohortColumns.ts
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export enum DatasetCohortColumns {
-  Index = "Index",
-  Dataset = "Data",
-  PredictedY = "Predicted Y",
-  TrueY = "True Y",
-  ClassificationError = "Classification outcome",
-  RegressionError = "Regression error",
-  ProbabilityY = "Probability Y"
-}
