+++ libs/localization/src/lib/en.json
@@ -1177,10 +1177,7 @@
       "selectionRate": "Selection rate: {0}",
       "mae": "Mean absolute error: {0}",
       "f1Score": "F1 score: {0}",
-      "samples": "Sample size: {0}",
-      "meanAveragePrecision": "Mean average precision: {0}",
-      "averagePrecision": "Average precision: {0}",
-      "averageRecall": "Average recall: {0}"
+      "samples": "Sample size: {0}"
     },
     "ValidationErrors": {
       "_inconsistentDimensions.comment": "Raise warning if arguments passed in have different sizes, listing dimensions of both mismatching pieces.",
@@ -1739,18 +1736,6 @@
         "meanPrediction": {
           "name": "Mean prediction",
           "description": "The average of all predictions."
-        },
-        "meanAveragePrecision": {
-          "name": "Mean Average Precision score",
-          "description": "Mean Average Precision for Object Detection models."
-        },
-        "averagePrecision": {
-          "name": "Average Precision score",
-          "description": "Average Precision for a chosen class for Object Detection models."
-        },
-        "averageRecall": {
-          "name": "Average Recall score",
-          "description": "Average Recall for a chosen class for Object Detection models."
         },
         "fairnessMetricDifference": "Difference",
         "fairnessMetricRatio": "Ratio"
