+++ libs/model-assessment/src/index.ts
@@ -3,4 +3,5 @@
 
 export * from "./lib/ModelAssessmentDashboard/ModelAssessmentDashboard";
 export * from "./lib/ModelAssessmentDashboard/ModelAssessmentDashboardProps";
+export * from "./lib/ModelAssessmentDashboard/FeatureFlights";
 export * from "./lib/ModelAssessmentDashboard/Controls/ModelOverview/StatsTableUtils";
