+++ libs/model-assessment/src/lib/ModelAssessmentDashboard/Controls/ModelOverview/ModelOverview.tsx
@@ -23,7 +23,6 @@ import {
   JointDataset,
   ModelTypes,
   MultilabelMetrics,
-  ObjectDetectionMetrics,
   FluentUIStyles,
   MulticlassClassificationMetrics,
   ErrorCohort,
@@ -137,14 +136,6 @@ export class ModelOverview extends React.Component<
       defaultSelectedMetrics = [
         MultilabelMetrics.ExactMatchRatio,
         MultilabelMetrics.HammingScore
-      ];
-    } else if (
-      this.context.dataset.task_type === DatasetTaskType.ObjectDetection
-    ) {
-      defaultSelectedMetrics = [
-        ObjectDetectionMetrics.MeanAveragePrecision,
-        ObjectDetectionMetrics.AveragePrecision,
-        ObjectDetectionMetrics.AverageRecall
       ];
     } else {
       // task_type === "regression"
